Выносит createSelectedOffersTemplate из утилит в темлпейт
