+++ src/utils/point.js
@@ -104,20 +104,6 @@ const getSelectedOffers = (typeOffers, pointOffersIds) => {
   return typeOffers.filter((offer) => pointOffersIds.includes(offer.id));
 };
 
-const createSelectedOffersTemplate = (hasSelectedOffers, selectedOffers) => (
-  `${hasSelectedOffers ? `
-          <ul class="event__selected-offers">
-              ${selectedOffers.map(({ title, price }) =>
-    `<li class="event__offer">
-              <span class="event__offer-title">${title}</span>
-              +€
-              <span class="event__offer-price">${price}</span>
-            </li>`).join('')}
-          </ul>`
-    : ''}`
-);
-
-
 export {
   humanizePointDateTime,
   humanizePointDateDate,
@@ -137,6 +123,5 @@ export {
   getDestinationPhotos,
   getDestinationObject,
   getSelectedOffers,
-  createSelectedOffersTemplate,
   toUpperCaseFirstLetter,
 };
